Full inversion
