+++ src/VisualStudio/Core/Def/ProjectSystem/MetadataReferences/VisualStudioMetadataReferenceManager.Factory.cs
@@ -29,7 +29,7 @@ namespace Microsoft.VisualStudio.LanguageServices.Implementation.ProjectSystem
             if (_singleton == null)
             {
                 // If we're in VS we know we must be able to get a TemporaryStorageService
-                var temporaryStorage = (TemporaryStorageServiceFactory.TemporaryStorageService)workspaceServices.GetRequiredService<ITemporaryStorageService>();
+                var temporaryStorage = (TemporaryStorageService)workspaceServices.GetRequiredService<ITemporaryStorageService>();
                 Interlocked.CompareExchange(ref _singleton, new VisualStudioMetadataReferenceManager(_serviceProvider, temporaryStorage), null);
             }
